Minor Code Updates - 252s-dev

+++ Skill/Db/EntryMap.php
@@ -31,7 +31,7 @@ class EntryMap extends \App\Db\Mapper
             $this->dbMap->addPropertyMap(new Db\Text('title'));
             $this->dbMap->addPropertyMap(new Db\Text('assessor'));
             $this->dbMap->addPropertyMap(new Db\Integer('absent'));
-            $this->dbMap->addPropertyMap(new Db\Integer('confirm'));
+            $this->dbMap->addPropertyMap(new Db\Text('confirm'));
             $this->dbMap->addPropertyMap(new Db\Text('status'));
             $this->dbMap->addPropertyMap(new Db\Text('notes'));
             $this->dbMap->addPropertyMap(new Db\Date('modified'));
@@ -54,8 +54,8 @@ class EntryMap extends \App\Db\Mapper
             $this->formMap->addPropertyMap(new Form\Integer('placementId'));
             $this->formMap->addPropertyMap(new Form\Text('title'));
             $this->formMap->addPropertyMap(new Form\Text('assessor'));
-            $this->formMap->addPropertyMap(new Form\Text('absent'));
-            $this->formMap->addPropertyMap(new Form\Integer('confirm'));
+            $this->formMap->addPropertyMap(new Form\Integer('absent'));
+            $this->formMap->addPropertyMap(new Form\Text('confirm'));
             $this->formMap->addPropertyMap(new Form\Text('status'));
             $this->formMap->addPropertyMap(new Form\Text('notes'));
         }
@@ -128,7 +128,7 @@ class EntryMap extends \App\Db\Mapper
         }
 
         $res = $this->selectFrom($from, $where, $tool);
-        
+        //vd($from, $where);
         return $res;
     }
 
@@ -179,11 +179,16 @@ class EntryMap extends \App\Db\Mapper
      * @param int $entryId
      * @param int $itemId
      */
-    public function removeValue($entryId, $itemId)
+    public function removeValue($entryId, $itemId = null)
     {
-        $st = $this->getDb()->prepare('DELETE FROM skill_value WHERE entry_id = ? AND item_id = ?');
+        $st = $this->getDb()->prepare('DELETE FROM skill_value WHERE entry_id = ?');
         $st->bindParam(1, $entryId);
-        $st->bindParam(2, $itemId);
+        if ($itemId !== null) {
+            $st = $this->getDb()->prepare('DELETE FROM skill_value WHERE entry_id = ? AND item_id = ?');
+            $st->bindParam(1, $entryId);
+            $st->bindParam(2, $itemId);
+        }
+
         $st->execute();
     }
